fix: do not call admission webhook update on CRD deletion
see https://gravitee.atlassian.net/browse/GKO-1208

+++ internal/admission/api/v2/validate.go
@@ -92,6 +92,10 @@ func validateUpdate(
 		return errs
 	}
 
+	if newApi.IsBeingDeleted() {
+		return errs
+	}
+
 	base.DeleteDefinitionConfigMapIfNeeded(ctx, oldApi, newApi)
 
 	errs.Add(base.ValidateSubscribedPlans(ctx, oldApi, newApi, search.ApiV2SubsField))
